refactor: RegionStore, RegionRedisRepository 삭제(#43, #47)
* UserInfoResponse -> UserDetailResponse 이름 변경
* RegionResponse -> RegionInfoResponse 이름 변경
* UserController: API 버전 업 및 사용자 개인 정보 조회 API 변경 (/info -> /me)
