+++ src/main/java/jeonseguard/backend/region/infrastructure/repository/RegionRepositoryImpl.java
@@ -1,4 +1,4 @@
-package jeonseguard.backend.region.infrastructure;
+package jeonseguard.backend.region.infrastructure.repository;
 
 import jeonseguard.backend.region.domain.entity.Region;
 import jeonseguard.backend.region.domain.repository.RegionRepository;
